consolidate exclusion dialogs to facilitate edits
