+++ src/web/components/ExclusionZonesPanel/Dialogs/PlacementErrorDialog/PlacementErrorDialog.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
-import { JaiaContext, JaiaDispatchContext } from "../../context/JaiaContext";
-import { JaiaActions } from "../../context/jaia-actions";
+import { JaiaContext, JaiaDispatchContext } from "../../../../context/JaiaContext";
+import { JaiaActions } from "../../../../context/jaia-actions";
 
 /**
  * Modal shown when a waypoint placement is blocked because the location falls
